refactor : Constants 리팩토링으로 인한 에러 해결
